upgrade migration & types
