+++ crml/nft/src/migration.rs
@@ -1,18 +1,22 @@
 #[allow(dead_code)]
-pub mod v1_storage {
-	use crate::{CollectionId, Config, ListingId, MultiCurrency, RoyaltiesSchedule, SeriesId, TokenId};
-	use codec::{Decode, Encode};
-	use scale_info::TypeInfo;
-	use sp_std::prelude::*;
+use crate::pallet::{AuctionListing, FixedPriceListing, Listing, Listings, StorageVersion, CollectionId, Config, SeriesId, log};
+use crate::types::{ListingId, MultiCurrency, Releases, RoyaltiesSchedule};
+use cennznet_primitives::types::TokenId;
+use frame_support::{IterableStorageMap, StoragePrefixedMap, weights::Weight};
+use sp_runtime::traits::Zero;
+use sp_std::prelude::*;
 
-	#[derive(Decode, Encode, Debug, Clone, PartialEq, TypeInfo)]
+pub mod v1_storage {
+	use super::*;
+
+	#[derive(codec::Encode, codec::Decode, Debug, Clone, PartialEq, scale_info::TypeInfo)]
 	pub enum MetadataBaseURI {
 		Ipfs,
 		Https(Vec<u8>),
 	}
 
 	/// A type of NFT sale listing
-	#[derive(Debug, Clone, Encode, Decode, PartialEq, Eq, TypeInfo)]
+	#[derive(Debug, Clone, codec::Encode, codec::Decode, PartialEq, Eq, scale_info::TypeInfo)]
 	#[scale_info(skip_type_params(T))]
 	pub enum Listing<T: Config> {
 		FixedPrice(FixedPriceListing<T>),
@@ -20,7 +24,7 @@ pub mod v1_storage {
 	}
 
 	/// Information about an auction listing v1
-	#[derive(Debug, Clone, Encode, Decode, PartialEq, Eq, TypeInfo)]
+	#[derive(Debug, Clone, codec::Encode, codec::Decode, PartialEq, Eq, scale_info::TypeInfo)]
 	#[scale_info(skip_type_params(T))]
 	pub struct AuctionListing<T: Config> {
 		/// The asset to allow bids with
@@ -38,7 +42,7 @@ pub mod v1_storage {
 	}
 
 	/// Information about a fixed price listing v1
-	#[derive(Debug, Clone, Encode, Decode, PartialEq, Eq, TypeInfo)]
+	#[derive(Debug, Clone, codec::Encode, codec::Decode, PartialEq, Eq, scale_info::TypeInfo)]
 	#[scale_info(skip_type_params(T))]
 	pub struct FixedPriceListing<T: Config> {
 		/// The asset to allow bids with
@@ -57,7 +61,6 @@ pub mod v1_storage {
 		pub royalties_schedule: RoyaltiesSchedule<T::AccountId>,
 	}
 
-	pub struct Module<T>(sp_std::marker::PhantomData<T>);
 	frame_support::decl_storage! {
 		trait Store for Module<T: Config> as Nft {
 			pub IsSingleIssue get(fn is_single_issue): double_map hasher(twox_64_concat) CollectionId, hasher(twox_64_concat) SeriesId => bool;
@@ -65,5 +68,64 @@ pub mod v1_storage {
 			pub SeriesMetadataURI get(fn series_metadata_uri): double_map hasher(twox_64_concat) CollectionId, hasher(twox_64_concat) SeriesId => Option<Vec<u8>>;
 			pub Listings get(fn listings): map hasher(twox_64_concat) ListingId => Option<Listing<T>>;
 		}
+	}
+
+	frame_support::decl_module! {
+		pub struct Module<T: Config> for enum Call where origin: T::Origin { }
+	}
+}
+
+pub fn migrate_to_v2<T: Config>() -> Weight {
+	if StorageVersion::<T>::get() == Releases::V1 as u32 {
+		StorageVersion::<T>::put(Releases::V2 as u32);
+		v1_storage::CollectionMetadataURI::remove_all(None);
+		v1_storage::IsSingleIssue::remove_all(None);
+
+		let listings: Vec<(ListingId, v1_storage::Listing<T>)> = v1_storage::Listings::<T>::iter().collect();
+		let weight = listings.len() as Weight;
+		for (listing_id, listing) in listings {
+			let listing_migrated = match listing {
+				v1_storage::Listing::<T>::FixedPrice(v1_storage::FixedPriceListing {
+					fixed_price,
+					close,
+					payment_asset,
+					seller,
+					buyer,
+					tokens,
+					royalties_schedule,
+				}) => Listing::<T>::FixedPrice(FixedPriceListing {
+					fixed_price,
+					close,
+					payment_asset,
+					seller,
+					buyer,
+					tokens,
+					royalties_schedule,
+					marketplace_id: None,
+				}),
+				v1_storage::Listing::<T>::Auction(v1_storage::AuctionListing {
+					reserve_price,
+					close,
+					payment_asset,
+					seller,
+					tokens,
+					royalties_schedule,
+				}) => Listing::<T>::Auction(AuctionListing {
+					reserve_price,
+					close,
+					payment_asset,
+					seller,
+					tokens,
+					royalties_schedule,
+					marketplace_id: None,
+				}),
+			};
+			Listings::insert(listing_id, listing_migrated);
+		}
+
+		log!(warn, "🃏 listings migrated");
+		return 6_000_000 as Weight + weight * 100_000;
+	} else {
+		Zero::zero()
 	}
 }
